Refactor to use mybatis generator

+++ src/main/java/com/sundingyi/forum/controller/PublishController.java
@@ -34,7 +34,7 @@ public class PublishController {
     public String doPublish(@RequestParam("title") String title,
                             @RequestParam("description") String description,
                             @RequestParam("tag") String tag,
-                            @RequestParam(name = "id", required = false) Long id,
+                            @RequestParam(name = "id", required = false) Integer id,
                             HttpServletRequest httpServletRequest,
                             Model model) {
         model.addAttribute("title", title);
@@ -61,7 +61,7 @@ public class PublishController {
     }
     
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Long id,
+    public String edit(@PathVariable("id") Integer id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
